requirements.txt
langchain==0.1.0
openai==0.28.1
faiss-cpu==1.7.4
streamlit==1.32.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
PyGithub==2.2.0
tiktoken==0.6.0
python-multipart==0.0.9 